libskysense-local: do not read from serial port in rsp_num < 0
That is a hack and made in order to support RESET command, which
indeed (surprise!) resets the device.  When device is being reset
tty is closed and stupid libserialport receives 0 from read() and
does not handle it properly (simply spins and consumes CPU).

To avoid this improper behaviour do not read from a port only and
only when rsp_num < 0, i.e. for RESET command.