skybroker: wait for ifaces up at early boot start
That is needed if daemon is started on boot before network is up.
And yes, I know that systemd service should have

 After=network-online.target
 Wants=network-online.target

and actually it has, but it does not work.

I am tired fighting with systemd and all this admin stuff,
so just wait for reasonable time and continue.