libskysense-remote,skybroker,skyserver: set ZMQ_LINGER for each ZMQ socket
ZMQ leaks socket fds and memory even zmq_close() is called (see
ZMQ_LINGER manual).  If connection can't be established and zmq_send()
is called socket will be freed *only* when zmq context is terminated,
thus for loops in skyserver and skybroker we have a huge leak of fds.

Set ZMQ_LINGER for each socket and be happy.