libskysense-local: lock UART port before probing
There is unresolved bug, when daemon and client tool both
access same hardware and UART responses mix up.

Once I could reproduce the problem but never investigated
it thorougly. The problem was also reported by the customer.

The only weak place in the code is probing. So lock the tty
even before probing.