libskycharge,remote,server: make subscriptions secure
Patch introduces an API call which returns a subscription
token, which is device-host dependent. Since each call
(including the new one) passes complete authorization
by the broker using user/device uuid pair it is difficult
to compromise the token without completing the authorization.

Previously it was enough to know the device-uuid to subscribe
on charging events.

Token is built using the /etc/machine-id of the BBB and
the device-uuid, so that device-uuid is not leaked.