skyclient: verify command line arguments before making a connection to a device
This is needed to verify command line arguments, for example
this should lead to a syntax error:

   skycharge-cli set-dev-params detect-mode plc

because should be specified as the following:

   skycharge-cli set-dev-params detect-mode:plc

but in the first variant 'plc' is treated as a hostname,
so the connection hangs because zmq tries to establish
a connection to the 'plc' host.

So in this patch parameters are parsed before any connection
attempt is made.