libskycharge-local: pass va_list by pointer, not by value
On amd64 this is not a problem, since va_list is already defined
as a pointer, but on arm (bbb) this defined as a struct, which
is copied on the stack if it is passed by a value. In order to
iterate over the arguments in the skycmd_arg_copy() function
we need a pointer.

This fixes HW1 set_param() call.